Создала объект и функции к нему. И в стратегии добавила вывод процента

+++ folderstrategy.h
@@ -8,8 +8,8 @@ public:
     FolderSrtategy(){};
     void DoStrategy(QString&  path);
 private:
-    quint32 sizeFolder(QString  path);
-    quint32 sizeListFolder(QString  path);
+    qint64 sizeFolder(QString  path);
+    qint64 sizeListFolder(QString  path);
 };
 
 #endif // FOLDERSTRATEGY_H
